Use a PS2 keyboard to be a parallel keyboard
Often for old computers on needs a parallel output type computer. This code will use your Blue Pill into a converter that read the keys from the keyboard and convert them to parallel output. Only normal keys are translated. No arrow, function or number pad are translated. It does handle shift lock, shift and control as one of these older keyboards would do. Of course one could add any operation on desired. Tables for the PS2 are on the web or you can just read them with this code. It is not turnkeyed. That is up to you.