List file for current EFORTH
I hot patch code in binary to the F103 so it has not been reassembled. I have attempted to keep the list file up to date with what the instructions are in the code. There are a number of unused locations around the TURNKEY code area. These could have any values in them. Some of the pointers to locations show at 20000000 while others have values. The original assemble didn't link these values. Those that are shown were entered by me. I most cases one can use a find in an editor or just talk to the EFORTH to see them.