INTHEX.F is loaded on the F103 to save a modified image.
In the TALKF105 there is a LogHex word. It is run on the PC to upload the data from the F103 that is running INTHEX ( adr cnt - ). Make sure to set CP, LAST and CURRENT in the boot user table to the desired values before running. The boot table is at 200001B4 and has the same offset as the running user table at 20004F00.